// apps/api/src/dto/event.dto.d.ts
export declare class EventDto {
    azureStorageContainer: string;
    minAge: number;
    maxAge: number;
    minGuardianAge: number;
    maxRegistration: number;
    maxVoucher: number;
    eventBeginDate: Date;
    registrationOpenDate: Date;
    registrationClosedDate: Date;
    projectClosedDate: Date;
    officialStartDate: Date;
    eventEndDate: Date;
    event_title: string;
    maxFileSize: number;
}